Separate argument parsing into its own file
